miko/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-telegram-bot==21.6
httpx==0.27.2
mem0ai==0.1.29
qdrant-client==1.11.3
pydantic==2.9.2
